apps/api/makeup/requirements.txt
gradio==4.44.0
huggingface_hub==0.25.2
mediapipe==0.10.14

numpy==1.26.4
opencv-python-headless==4.9.0.80

torch==2.1.2
torchvision==0.16.2
onnxruntime==1.17.3

Pillow==10.1.0
protobuf>=4.25.3,<5
requests==2.31.0
typing-extensions==4.9.0
websockets==11.0.3
